Converted files with function to classes and implemented layout for flask incorpotation
